Push/Pull
Can now grab crates and push/pull them. When letting go you have to hold
x until you're a small distance away from the box to fully let go.
However there are bugs. Sometimes on the first grab the box will start
to slide uncontrollably in the first direction. This can be stopped by
pressing x again and "dropping" the box, but I'm not sure what causes it
in the first place. I'll take a look into possibly setting a max
distance before the box is forcefully dropped. Also if you move the
player to the left side of the box and grab it there is a bug where it
only seems to register the player's velocity moving to the left, so it
can't move right