Push/Pull
Finished push/pull for boxes. Works on both sides, no random sliding,
and the playing doesn't have to hold down the action button to release
the crate. There is a cooldown of when the player can put down the crate
after picking it up (and vice versa) that is currently set to 1 second,
but can be changed with a simple number change at line 104 to whatever
feels best. Otherwise the player simply presses 'X' to pick up and 'X'
to let go. Will work on push/pull ladders once Harvey has them
working/something for me to work with.