Motion Platform
This is Josh's motion platform.  The pull request was a little messed up
so I had to reassemble the parts in Unity manually since we're on a
tight schedule atm.
I just (at least temporarily) fixed the child to abnormal scale bug by
giving the platform a parent that's scale (1,1,1) and holds the
components.
- Steve