Push/Pull Update
I manually merged in Branden's push/pull update.  I made a couple
changes (let me know how you feel about them) to get it to work again.
One thing we need to fix is that if you're directly up against it and
try to push, it doesn't work because you're colliding with it and it's
inheriting your 0 velocity.  One fix would be to somehow make it so
Anais and the box don't collide when in push/pull.